Revert "Revert "- change the reg. policy to force user to put image""
This reverts commit 598c4904a36a11aa8a1eab03e10b1519a16fefbf.

+++ lib/pages/others/map.dart
@@ -643,9 +643,9 @@ class _MapPageState extends State<MapPage> {
                                     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
                                     print(meter);
                                     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
-                                    formdata.add("distance", meter);
+                                    formdata.add("distance", meter.toInt());
                                     formdata.add("steps", _polylineIdCounter);
-                                    formdata.add("calories", _polylineIdCounter*0.0512);
+                                    formdata.add("calories", (_polylineIdCounter*0.0512).toInt());
                                     print(
                                         '!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!');
                                     print(formdata);
